admin/requirements.txt
Flask==3.0.0
Flask-Admin==1.6.1
Flask-Login==0.6.3
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0